# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: Coupons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"
  /coupons/{couponId}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - in: path
          name: couponId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coupon deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Notifications
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        maxUses:
          type: integer
          nullable: true
        usedCount:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
        courseId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, code, discountType, discountValue, usedCount, isActive, createdAt]
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        maxUses:
          type: integer
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        courseId:
          type: integer
          nullable: true
      required: [code, discountType, discountValue]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        courseId:
          type: integer
          nullable: true
        bundleId:
          type: integer
          nullable: true
      required: [code]
    CouponValidation:
      type: object
      properties:
        valid:
          type: boolean
        discountType:
          type: string
          nullable: true
        discountValue:
          type: number
          nullable: true
        finalPrice:
          type: number
          nullable: true
        message:
          type: string
      required: [valid, message]

    # Notifications
...