# lib/api-zod/src/generated/types/validateCouponBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ValidateCouponBody {
  code: string;
  courseId?: number | null;
  bundleId?: number | null;
}